requirements.txt
google-generativeai==0.5.0
streamlit==1.33.0